Style fixes; add two-pointer & search matrix
Apply small style and safety improvements (use static_cast for sizes/ceil, add const qualifiers, make mid/previousDay const, clean up formatting and remove example comment in ValidParentheses). Add new algorithm implementations: Binary-Search/SearchMatrix, Two-Pointer/ContainerWithMostWater, and Two-Pointer/ThreeSum (with copyright headers). Update CMakeLists.txt to include the new source files.

+++ Array-Hashing/ProductExceptSelf.cpp
@@ -33,7 +33,7 @@
 
 std::vector<int> productExceptSelf(const std::vector<int>& nums)
 {
-    int n = nums.size();
+    int n = static_cast<int>(nums.size());
 
     std::vector<int> result(n, 1);
